app/requirements.txt
kafka-python==2.0.2
Faker==9.5.0